Import interfaces

+++ Main.java
@@ -3,8 +3,8 @@ public class Main {
         Calculator calc = Calculator.instance.get();
         int a = calc.plus.apply(1, 2);
         int b = calc.minus.apply(1, 1);
-        // При делении на 0 выдаст ошибку, можно добавить проверку или перевести операцию деления в double
-        double c = calc.divide.apply(a, b);
+        // При делении на 0 выдаст ошибку
+        int c = calc.divide.apply(a, b);
         calc.println.accept(c);
     }
 }
